requirements.txt
jsonschema==3.2.0
PyYAML==5.3
munch==2.5.0
pytest==5.3